requirements.txt
flask~=1.1.2
line-bot-sdk
configparser~=5.0.0
django
gunicorn
django-heroku
requests
werkzeug>=0.15.4

apscheduler~=3.6.3
argparse~=1.4.0
pytz~=2020.1